requirements.txt
pymupdf4llm
pymupdf
langchain
langchain_text_splitters
boto3
Pillow
transformers